\name{Athaliana_ODE}
\alias{Athaliana_ODE}
\docType{data}
\title{Gene expression time series generated with ODE model}
\description{
This data set was generated using a circadian clock regulatory network ODE model. The data was simulated using COPASI, subsampled to produce 
hourly data and after this the log was calculated.
}
\usage{data(Athaliana_ODE)}
\format{A matrix containing genes in rows and (1h) time points in columns.}
\references{
Hoops, S., Sahle, S., Gauges, R., Lee, C., Pahle, J., Simus, N., Singhal, M., Xu, L., Mendes, P. and Kummer, U.
(2006) COPASI a COmplex PAthway SImulator. Bioinformatics, 22, 3067-3074.


Locke, J.C.W., Kozma-Bognar, L., Gould, P.D., Feher, B., Kevei, E., Nagy, F., Turner, M.S., Hall, A. and Millar,  A.J. (2006) 
Experimental validation of a predicted feedback loop in the multi-oscillator clock of Arabidopsis thaliana. Molecular Systems Biology
}
\keyword{datasets}
